move : 장비, 악세서리가 아닌 밑 추가 부분은 따로 코드 정리
작업 완료
팔찌 부분이 많아서 Extras로 뺌

작업 예정
Merge

장비
(나중에) 팔찌 상중하 옵션별로 색부여

보석
보석 Detail 글씨 크기 조절해 볼 생각

스킬
보석에서 mapping한 정보가 있다면 그것으로 보석 mapping, detail 손보기
간단설명에서 룬 TextChip에 (전설) -> 이걸 배경색으로 변경
